Calculate number of significant variables for a set of pvalues thresholds

Calculate number of significant variables for a set of pvalues thresholds

nsignif.DEswan(res.DEswan.wide, thresholds)

Arguments

res.DEswan.wide:
a data Frame - pvalues or qvalues from DEswan in a wide format

thresholds:
a Vector - p or q-values thresholds. Default is c(0.05,1e-2,1e-3,1e-4,1e-5)

Examples

res.DEswan=DEswan(data.df = agingplasmaproteome[,-c(1:3)],
qt = agingplasmaproteome[,1],
window.center = seq(40,100,10),
buckets.size = 10,
covariates = agingplasmaproteome[,c(2:3)])

#> [1] "window.center  1/7"
#> [1] "window.center  2/7"
#> [1] "window.center  3/7"
#> [1] "window.center  4/7"
#> [1] "window.center  5/7"
#> [1] "window.center  6/7"
#> [1] "window.center  7/7"

res.DEswan.wide.p=reshape.DEswan(res.DEswan,parameter = 1,factor = "qt")
res.DEswan.wide.p.signif=nsignif.DEswan(res.DEswan.wide.p)
head(res.DEswan.wide.p.signif)

#>       X40 X50 X60 X70 X80 X90 X100
#> 0.05   65  54  71 135 299 128  140
#> 0.01   25  15  14  33 157  37   60
#> 0.001   5   4   2   7  44   6   23
#> 1e-04   1   1   0   2  12   0   12
#> 1e-05   0   0   0   1   6   0    5